Crate whitenoise_validator

The Whitenoise rust validator contains methods for evaluating and constructing differentially private analyses.

The validator defines a set of statically checkable properties that are necessary for a differentially private analysis, and then checks that the submitted analysis satisfies the properties.

The validator also takes simple components from the Whitenoise runtime and combines them into more complex mechanisms.

Modules

base
Core data structures

bindings
Shorthand interface for building differentially private analyses.

components
Component trait implementations

docs
Standalone documentation files

ffi
Foreign function interfaces

proto

utilities

Functions

accuracy_to_privacy_usage
Estimate the privacy usage necessary to bound accuracy to a given value.

compute_privacy_usage
Compute overall privacy usage of an analysis.

expand_component
Expand a component that may be representable as smaller components, and propagate its properties.

generate_report
Generate a json string with a summary/report of the Analysis and Release

get_properties
Retrieve the static properties from every reachable node on the graph.

privacy_usage_to_accuracy
Estimate the accuracy of the release of a component, based on a privacy usage.

validate_analysis
Validate if an analysis is well-formed.