Module whitenoise_validator::base

Core logic and data structures

Structs

AggregatorProperties
Properties of an aggregation applied to a Value.

ArrayNDProperties
Derived properties for the universal ArrayND.

HashmapProperties
Derived properties for the universal Hashmap.

NatureCategorical

NatureContinuous

Vector2DJaggedProperties
Derived properties for the universal Vector2DJagged.

Enums

ArrayND
The universal n-dimensional array representation.

DataType
Fundamental data types for ArrayNDs and Vector2DJagged Values.

Hashmap
The universal hash-map representation.

Nature

Sensitivity
Accepted spaces for sensitivity to be computed within.

Value
The universal data representation.

ValueProperties
Derived properties for the universal value.

Vector1DNull

Vector1D

Vector2DJagged
The universal jagged array representation.

Functions

compute_privacy_usage

expand_component

generate_report

get_argument
Retrieve the specified Value from the arguments to a component.

get_component_privacy_usage

get_input_arguments
Retrieve the Values for each of the arguments of a component from the Release.

get_input_properties
Retrieve the ValueProperties for each of the arguments of a component from the Release.

get_traversal
Given a computation graph, return an ordering of nodes that ensures all dependencies of any node have been visited

privacy_usage_reducer

propagate_properties
Given an analysis and release, attempt to propagate properties across the entire computation graph.

Type Definitions

NodeProperties

Release
A release consists of Values for each node id.