Enum whitenoise_validator::base::Value

pub enum Value {
    Array(Array),
    Hashmap(Hashmap<Value>),
    Jagged(Jagged),
}

The universal data representation.

Arguments to components are hash-maps of Value and the result of a component is a Value. The Value is also used in the validator for public arguments.

The Value has a one-to-one mapping to a protobuf Value.

Components unwrap arguments into more granular types, like ndarray::Array1, run a computation, and then repackage the result back into a Value.

Variants

Array(Array)

An arbitrary-dimensional homogeneously typed array

Hashmap(Hashmap<Value>)

A hash-map, where the keys are enum-typed and the values are of type Value

Jagged(Jagged)

A 2D homogeneously typed matrix, where the columns may be unknown and the column lengths may be inconsistent

Methods

impl Value

pub fn array(&self) -> Result<&Array>

Retrieve an Array from a Value, assuming the Value contains an Array

pub fn jagged(&self) -> Result<&Jagged>

Retrieve Jagged from a Value, assuming the Value contains Jagged

pub fn first_f64(&self) -> Result<f64>

Retrieve the first f64 from a Value, assuming a Value contains an ArrayND of type f64

pub fn first_i64(&self) -> Result<i64>

Retrieve the first i64 from a Value, assuming a Value contains an ArrayND of type i64

pub fn first_string(&self) -> Result<String>

Retrieve the first String from a Value, assuming a Value contains an ArrayND of type String

pub fn first_bool(&self) -> Result<bool>

Retrieve the first bool from a Value, assuming a Value contains an ArrayND of type bool