Struct whitenoise_validator::base::ArrayProperties

pub struct ArrayProperties {
    pub num_records: Option<i64>,
    pub num_columns: Option<i64>,
    pub nullity: bool,
    pub releasable: bool,
    pub c_stability: Vec<f64>,
    pub aggregator: Option<AggregatorProperties>,
    pub nature: Option<Nature>,
    pub data_type: DataType,
    pub dataset_id: Option<i64>,
    pub is_not_empty: bool,
    pub dimensionality: u32,
}

Derived properties for the universal ArrayND.

The ArrayNDProperties has a one-to-one mapping to a protobuf ArrayNDProperties.

Fields
num_records: Option<i64>

Defined if the number of records is known statically (set by the resize component)

num_columns: Option<i64>nullity: bool

true if the data may contain null values

releasable: bool

set to true by the mechanisms. Acts as a filter on the values in the release

c_stability: Vec<f64>

amplification of privacy usage by unstable data transformations, or possibility of duplicated records

aggregator: Option<AggregatorProperties>

set when data is aggregated, used to help compute sensitivity from the mechanisms

nature: Option<Nature>

either min/max or categories

data_type: DataType

f64, i64, bool, String

dataset_id: Option<i64>

index of last Materialize or Filter node, where dataset was created used to determine if arrays are conformable even when N is not known

is_not_empty: bool

true if the array may not be length zero

dimensionality: u32

number of axes in the array

Methods

impl ArrayProperties

pub fn lower_f64_option(&self) -> Result<Vec<Option<f64>>>

pub fn lower_f64(&self) -> Result<Vec<f64>>

pub fn upper_f64_option(&self) -> Result<Vec<Option<f64>>>

pub fn upper_f64(&self) -> Result<Vec<f64>>

pub fn lower_i64_option(&self) -> Result<Vec<Option<i64>>>

pub fn lower_i64(&self) -> Result<Vec<i64>>

pub fn upper_i64_option(&self) -> Result<Vec<Option<i64>>>

pub fn upper_i64(&self) -> Result<Vec<i64>>

pub fn categories(&self) -> Result<Jagged>

pub fn assert_categorical(&self) -> Result<()>

pub fn assert_non_null(&self) -> Result<()>

pub fn assert_is_not_empty(&self) -> Result<()>

pub fn assert_is_releasable(&self) -> Result<()>

pub fn num_columns(&self) -> Result<i64>

pub fn num_records(&self) -> Result<i64>

pub fn assert_is_not_aggregated(&self) -> Result<()>