Module whitenoise_validator::components

Component trait implementations

Each component represents an abstract computation. Example components are Materialize for loading a dataframe, Index for retrieving specific columns from the dataframe, Mean for aggregating data, LaplaceMechanism for privatizing data, etc.

There are a set of possible behaviours each component may implement. Each behavior corresponds to a trait. A listing of traits is at the bottom of the page.

Implementations of the traits are distributed among the module files.

Modules

index

Traits

Accuracy
Accuracy component trait

Aggregator
Aggregator Component trait

Component
Universal Component trait

Expandable
Expandable Component trait

Named
Named component trait

Report
Report component trait