Trait whitenoise_validator::components::Named

pub trait Named {
    fn get_names(
        &self, 
        public_arguments: &HashMap<String, Value>, 
        argument_variables: &HashMap<String, Vec<String>>, 
        release: &Option<&Value>
    ) -> Result<Vec<String>>;
}

Named component trait

Named components involve variables and keep track of the human readable names for these variables and may modify these variables names.

Required methods

fn get_names(
&self,
public_arguments: &HashMap<String, Value>,
argument_variables: &HashMap<String, Vec<String>>,
release: &Option<&Value>
) -> Result<Vec<String>>

Propagate the human readable names of the variables associated with this component

Loading content...