Trait whitenoise_validator::components::Utility

pub trait Utility {
    fn get_utility(
        &self, 
        privacy_definition: &PrivacyDefinition
    ) -> Result<Utility>;
}

Utility component trait

Components with utility implemented may be privatized with the exponential mechanism

Required methods

fn get_utility(&self, privacy_definition: &PrivacyDefinition) -> Result<Utility>

return a function represented in protobuf that computes the utility associated with this component

Loading content...