Module whitenoise_validator::docs::components

All of the components available in the library are listed below. The components may be strung together in arbitrary directed graphs (called analyses), and only verifiably DP analyses and data are released.

Component ID	Bindings Name	Inputs

Abs	abs	data

Add	add	left, right

Cast	cast	atomic_type, data, lower, true_label, upper

Clamp	clamp	categories, data, lower, null_value, upper

Count	count	data

Covariance	covariance	data, finite_sample_correction, left, right

Digitize	digitize	data, edges, inclusive_left, null_value

Divide	divide	left, right

DPCount	dp_count	data, enforce_constant_time, lower, mechanism, privacy_usage, upper

DPCovariance	dp_covariance	data, finite_sample_correction, left, mechanism, privacy_usage, right

DPHistogram	dp_histogram	categories, data, edges, enforce_constant_time, inclusive_left, lower, mechanism, null_value, privacy_usage, upper

DPMaximum	dp_maximum	data, mechanism, privacy_usage

DPMean	dp_mean	data, implementation, mechanism, privacy_usage

DPMedian	dp_median	data, interpolation, mechanism, privacy_usage

DPMinimum	dp_minimum	data, mechanism, privacy_usage

DPMomentRaw	dp_moment_raw	data, mechanism, order, privacy_usage

DPSum	dp_sum	data, lower, mechanism, privacy_usage, upper

DPVariance	dp_variance	data, finite_sample_correction, mechanism, privacy_usage

Equal	equal	left, right

Filter	filter	data, mask

GaussianMechanism	gaussian_mechanism	data, privacy_usage

GreaterThan	greater_than	left, right

Histogram	histogram	categories, data, edges, inclusive_left, null_value

Impute	impute	categories, data, distribution, lower, scale, shift, upper, weights

Index	index	columns, data

KthRawSampleMoment	kth_raw_sample_moment	data, k

LaplaceMechanism	laplace_mechanism	data, privacy_usage

LessThan	less_than	left, right

Literal	literal

Log	log	base, data

And	logical_and	left, right

Or	logical_or	left, right

Materialize	materialize	column_names, data_source, dataset_id, public, skip_row

Maximum	maximum	data

Mean	mean	data

Minimum	minimum	data

Modulo	modulo	left, right

Multiply	multiply	left, right

Negate	negate	data

Negative	negative	data

Partition	partition	by, data, num_partitions

Power	power	data, radical

Quantile	quantile	alpha, data, interpolation

Reshape	reshape	data, layout, shape, symmetric

Resize	resize	categories, data, distribution, lower, n, scale, shift, upper, weights

RowMax	row_max	left, right

RowMin	row_min	left, right

SimpleGeometricMechanism	simple_geometric_mechanism	data, enforce_constant_time, lower, privacy_usage, upper

Subtract	subtract	left, right

Sum	sum	data

ToBool	to_bool	data, true_label

ToFloat	to_float	data

ToInt	to_int	data, lower, upper

ToString	to_string	data

Variance	variance	data, finite_sample_correction