Function whitenoise_validator::validate_analysis

#[no_mangle]
pub extern "C" fn validate_analysis(
    request_ptr: *const u8, 
    request_length: i32
) -> ByteBuffer

Validate if an analysis is well-formed.

Checks that the graph is a DAG. Checks that static properties are met on all components.

Useful for static validation of an analysis. Since some components require public arguments, mechanisms that depend on other mechanisms cannot be verified until the components they depend on have been validated.

The system may also be run dynamically- prior to expanding each node, calling the expand_component endpoint will also validate the component being expanded. NOTE: Evaluating the graph dynamically opens up additional potential timing attacks.

Arguments

request_ptr - a pointer to an array containing the serialized protobuf of RequestValidateAnalysis

request_length - the length of the array

Returns

a ByteBuffer struct containing a pointer to and length of the serialized protobuf of proto::ResponseValidateAnalysis