Module whitenoise_validator::utilities

Modules

array

inference
Infer ValueProperties from a public Value

json
Representation for report/json summaries

serial
Serialization and deserialization between prost protobuf structs and internal representations

Functions

broadcast_ndarray

broadcast_privacy_usage

deduplicate

expand_mechanism
Utility function for building component expansions for dp mechanisms

get_argument
Retrieve the specified Value from the arguments to a component.

get_component_privacy_usage

get_input_arguments
Retrieve the Values for each of the arguments of a component from the Release.

get_input_properties
Retrieve the ValueProperties for each of the arguments of a component from the Release.

get_ith_release

get_traversal
Given a computation graph, return an ordering of nodes that ensures all dependencies of any node have been visited

is_conformable

privacy_usage_reducer

propagate_properties
Given an analysis and release, attempt to propagate properties across the entire computation graph.

standardize_float_argument