Trait whitenoise_validator::components::Expandable

pub trait Expandable {
    fn expand_component(
        &self, 
        privacy_definition: &Option<PrivacyDefinition>, 
        component: &Component, 
        public_arguments: &IndexMap<IndexKey, &Value>, 
        properties: &NodeProperties, 
        component_id: u32, 
        maximum_id: u32
    ) -> Result<ComponentExpansion>;
}

Expandable Component trait

When a component is expandable, it represents a higher order computation that may be expressed in multiple components that are more granular. Oftentimes Expandable components correspond to differentially private algorithms, that are represented in terms of an aggregation and a mechanism.

Required methods

fn expand_component(
&self,
privacy_definition: &Option<PrivacyDefinition>,
component: &Component,
public_arguments: &IndexMap<IndexKey, &Value>,
properties: &NodeProperties,
component_id: u32,
maximum_id: u32
) -> Result<ComponentExpansion>

Concrete implementation for an Expandable component that returns a patch that may be applied to a computation graph.

Arguments

self - the protobuf object corresponding to the prost protobuf struct

privacy_definition - definition of privacy to use when expanding. Some expansions are not valid under some privacy definitions

component - contains additional metadata about the argument node ids

properties - properties on the data supplied as arguments

component_id - the id of the node to expand. The final node in the returned patch must use this id.

maximum_id - the starting id for which additional nodes may be added to the graph without overwriting existing nodes

Returns

Sufficient information to patch the runtime with more granular steps. More documentation at ComponentExpansion.

Loading content...